restructurada el modulo vista
