+++ vista/User/UserFormPage.php
@@ -1,6 +1,6 @@
 <?php 
 
-	require_once("../modelo/formularios_modelo.php");
+	require_once("../../modelo/formularios_modelo.php");
 
  ?>
 <!DOCTYPE html>
@@ -11,14 +11,8 @@
     <meta name="description" content="">
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
-    <link rel="stylesheet" href="../css/bootstrap.min.css">
 	<title>Formularios</title>
-	<link rel="stylesheet" type="text/css" href="../css/style.css">
-	<link rel="stylesheet" href="../css/bootstrap-theme.min.css">
-    <link rel="stylesheet" href="../css/main.css">
-    <link href="https://fonts.googleapis.com/css?family=Roboto|Squada+One&display=swap" rel="stylesheet">  
-
-    <script src="../js/vendor/modernizr-2.8.3.min.js"></script>
+	<?php include_once('../Includes/header.php'); ?>
 </head>
 <body>
 	<nav class="navbar navbar-personalizado"> 
@@ -29,20 +23,20 @@
 					<span class="icon-bar app-bar"></span>
 					<span class="icon-bar app-bar"></span>
 				</button>
-				<a href="vista_usuarios.php" class="navbar-brand link-personalizado"><span class="glyphicon glyphicon-search"></span> Formularios Salutogenesis</a>
+				<a href="UserMainPage.php" class="navbar-brand link-personalizado"><span class="glyphicon glyphicon-search"></span> Formularios Salutogenesis</a>
 			</div>
 			<div class="collapse navbar-collapse" id="menu">
 				<ul class="nav navbar-nav navbar-right nav-personalizado">
-					<li><a href="vista_usuarios_formularios.php">Formularios</a></li>
-					<li><a href="vista_buscar.php">Busqueda</a></li>
+					<li><a href="UserFormPage.php">Formularios</a></li>
+					<li><a href="../vista_buscar.php">Busqueda</a></li>
 					<li class="dropdown"><a href="" class="dropdown-toggle" data-toggle="dropdown">
 							<div class="contenedo-usuario">
 								<img src=""><span class="glyphicon glyphicon-user"></span>
 							</div>
 							<ul class="dropdown-menu">
-								<li><a href="vista_acerca.php">Acerca de</a></li>
-								<li><a href="vista_perfil.php">Mi perfil</a></li>
-								<li><a href="../controlador/salir_controlador.php">Salir</a></li>
+								<li><a href="../Includes/about.php">Acerca de</a></li>
+								<li><a href="../Includes/profile.php">Mi perfil</a></li>
+								<li><a href="../../controlador/salir_controlador.php">Salir</a></li>
 							</ul>
 					</li>
 				</ul>
@@ -67,7 +61,7 @@
 
 					if(isset($_GET["pagina"])){
 								if($_GET["pagina"] == 1){
-									header("location: ../vista/vista_usuarios_formularios.php");
+									header("location: UserFormPage.php");
 								}else{
 									$pagina = $_GET["pagina"];
 								}
@@ -92,7 +86,7 @@
 				 	<td><?php echo $fila['idFormularios']; ?></td>
 				 	<td><?php echo $fila['nombre']; ?></td>
 				 	<td><?php echo $fila['descripcion']; ?></td>
-				 	<td><a href="vista_responder_formulario.php?id=<?php echo $fila['idFormularios'] ?>&num=<?php echo $fila['numeroPregunta']; ?>"><button class="btn btn-warning">Responder</button></a></td>
+				 	<td><a href="AnswerFormPage.php?id=<?php echo $fila['idFormularios'] ?>&num=<?php echo $fila['numeroPregunta']; ?>"><button class="btn btn-warning">Responder</button></a></td>
 				 </tr>
 				<?php } ?>
 			</table>
@@ -139,10 +133,6 @@
          </div>
        </div>   
 
- 	<script src="../js/main.js"></script>
-	<script src="//ajax.googleapis.com/ajax/libs/jquery/1.11.2/jquery.min.js"></script>
-    <script>window.jQuery || document.write('<script src="../js/vendor/jquery-1.11.2.min.js"><\/script>')</script>
-
-    <script src="../js/vendor/bootstrap.min.js"></script>
+	   <?php include_once('../Includes/footer.php'); ?>
 </body>
 </html>
